break up app into components
