+++ src/elements/logo/logo.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const Logo = styled.div`
+  width: 120px;
+  height: 31px;
+  background: rgba(255, 255, 255, 0.2);
+  margin: 16px 28px 16px 0;
+  float: left;
+`;
+
+
+const logo = () => (
+  <Logo />
+);
+
+export default logo;
